Move ChatTemplates to a module and add comments
